fix: 类型标注与异步生成器误用修复——fetch_data 改 async for、阈值 range 类型、序列化 float 收敛
- sync_tasks: async generator 的 fetch_data 被 await 后 for 迭代必抛 TypeError，改 async for
- training_data/base: check_threshold 支持 range 型 [min,max] 阈值并显式收敛类型
- reranker_metrics/embedding_metrics: np/round 结果显式 float，避免 JSON 序列化出 numpy 标量
- reranker_metrics: 修复中文引号相邻导致的字符串拼接错误（"...不够""难""" → "...不够难"）
- protocol_converter/evalscope_metrics: Optional 类型标注补齐

+++ backend/app/services/training_data/embedding_metrics.py
@@ -205,7 +205,7 @@ class EmbeddingCompletenessMetric(BaseTrainingDataMetric):
             completeness_checks.append(0.0)
             suggestions.append("缺少上下文数据")
 
-        score = np.mean(completeness_checks)
+        score = float(np.mean(completeness_checks))
 
         return TrainingDataMetricResult(
             score=score,
